redux 수정
