+++ src/redux/Containers/CounterContainers.js
@@ -1,25 +1,11 @@
 import React from "react";
 import Counter from "../components/Counter";
-import { useSelector, useDispatch } from "react-redux";
-import { increase, decrease } from "../module/counter";
+
 
 const CounterContainers = () => {
-  const number = useSelector((state) => state.counter);
-
-  const dispatch = useDispatch();
-
-  const onIncrease = () => {
-    dispatch(increase());
-  };
-  const onDecrease = () => {
-    dispatch(decrease());
-  };
+  
   return (
-    <Counter
-      number={number}
-      onIncrease={onIncrease}
-      onDecrease={onDecrease}
-    ></Counter>
+    <Counter />
   );
 };
